Moved gizmo into functions into cpp files
