+++ src/editor/translationgizmo.cpp
@@ -1,6 +1,42 @@
 #include "translationgizmo.h"
+#include <QOpenGLFunctions_3_2_Core>
+#include <qopenglshaderprogram.h>
+
 #include "irisgl/src/math/intersectionhelper.h"
 #include "irisgl/src/math/mathhelper.h"
+#include "irisgl/src/core/irisutils.h"
+#include "irisgl/src/scenegraph/scene.h"
+#include "irisgl/src/scenegraph/cameranode.h"
+#include "irisgl/src/graphics/graphicshelper.h"
+#include "uimanager.h"
+#include "../commands/transfrormscenenodecommand.h"
+#include "irisgl/src/math/mathhelper.h"
+
+TranslationHandle::TranslationHandle(Gizmo* gizmo, GizmoAxis axis)
+{
+	this->gizmo = gizmo;
+
+	switch (axis) {
+	case GizmoAxis::X:
+		handleExtent = QVector3D(1, 0, 0);
+		planes.append(QVector3D(0, 1, 0));
+		planes.append(QVector3D(0, 0, 1));
+		setHandleColor(QColor(255, 0, 0));
+		break;
+	case GizmoAxis::Y:
+		handleExtent = QVector3D(0, 1, 0);
+		planes.append(QVector3D(1, 0, 0));
+		planes.append(QVector3D(0, 0, 1));
+		setHandleColor(QColor(0, 255, 0));
+		break;
+	case GizmoAxis::Z:
+		handleExtent = QVector3D(0, 0, 1);
+		planes.append(QVector3D(1, 0, 0));
+		planes.append(QVector3D(0, 1, 0));
+		setHandleColor(QColor(0, 0, 255));
+		break;
+	}
+}
 
 bool TranslationHandle::isHit(QVector3D rayPos, QVector3D rayDir)
 {
@@ -87,45 +123,125 @@ QVector3D TranslationHandle::getHitPos(QVector3D rayPos, QVector3D rayDir)
 	return finalHitPos;
 }
 
-/*
-bool TranslationHandle::isHit(QVector3D rayPos, QVector3D rayDir)
+TranslationGizmo::TranslationGizmo() :
+	Gizmo()
 {
-	auto gizmoTrans = gizmo->getTransform();
+	handles[0] = new TranslationHandle(this, GizmoAxis::X);
+	handles[1] = new TranslationHandle(this, GizmoAxis::Y);
+	handles[2] = new TranslationHandle(this, GizmoAxis::Z);
 
-	// calculate world space position of the segment representing the handle
-	auto p1 = gizmoTrans * QVector3D(0, 0, 0);
-	auto q1 = gizmoTrans * handleExtent;
+	loadAssets();
+	//handle->setHandleColor(QColor(255, 255, 255));
 
-	auto p2 = rayPos;
-	auto q2 = rayPos + rayDir * 1000;
+	dragging = false;
+	draggedHandle = nullptr;
+}
 
-	float s, t;
-	QVector3D c1, c2;
-	auto dist = iris::MathHelper::closestPointBetweenSegments(p1, q1, p2, q2, s, t, c1, c2);
-	if (dist < 0.1f) {
-		return true;
+void TranslationGizmo::loadAssets()
+{
+	handleMeshes.append(iris::Mesh::loadMesh(IrisUtils::getAbsoluteAssetPath("app/models/axis_x.obj")));
+	handleMeshes.append(iris::Mesh::loadMesh(IrisUtils::getAbsoluteAssetPath("app/models/axis_y.obj")));
+	handleMeshes.append(iris::Mesh::loadMesh(IrisUtils::getAbsoluteAssetPath("app/models/axis_z.obj")));
+
+	shader = iris::GraphicsHelper::loadShader(
+		IrisUtils::getAbsoluteAssetPath("app/shaders/gizmo.vert"),
+		IrisUtils::getAbsoluteAssetPath("app/shaders/gizmo.frag"));
+}
+
+bool TranslationGizmo::isDragging()
+{
+	return dragging;
+}
+
+void TranslationGizmo::startDragging(QVector3D rayPos, QVector3D rayDir)
+{
+	//qDebug() << "drag starting";
+	draggedHandle = getHitHandle(rayPos, rayDir, hitPos);
+	if (draggedHandle == nullptr) {
+		dragging = false; // end dragging if no handle was actually hit
+		return;
+	}
+
+	nodeStartPos = selectedNode->getGlobalPosition();
+	dragging = true;
+}
+
+void TranslationGizmo::endDragging()
+{
+	dragging = false;
+}
+
+void TranslationGizmo::drag(QVector3D rayPos, QVector3D rayDir)
+{
+	//qDebug() << "dragging";
+	if (draggedHandle == nullptr) {
+		//dragging = false;
+		return;
+	}
+	//qDebug()<<"sliding";
+	auto slidingPos = draggedHandle->getHitPos(rayPos, rayDir);
+
+	// move node along line
+	// do snapping here as well
+	auto diff = slidingPos - hitPos;
+
+	selectedNode->setLocalPos(nodeStartPos + diff);
+	//selectedNode->setLocalScale(diff * 0.1f)
+}
+
+bool TranslationGizmo::isHit(QVector3D rayPos, QVector3D rayDir)
+{
+	for (auto i = 0; i< 3; i++) {
+		if (handles[i]->isHit(rayPos, rayDir))
+		{
+			//handle->setHandleColor(QColor(255, 255, 255));
+			return true;
+		}
 	}
 
 	return false;
 }
 
-QVector3D TranslationHandle::getHitPos(QVector3D rayPos, QVector3D rayDir)
+// returns hit position of the hit handle
+TranslationHandle* TranslationGizmo::getHitHandle(QVector3D rayPos, QVector3D rayDir, QVector3D& hitPos)
 {
-	auto gizmoTrans = gizmo->getTransform();
+	TranslationHandle* closestHandle = nullptr;
+	float closestDistance = 10000000;
 
-	// calculate world space position of the segment representing the handle
-	auto p1 = gizmoTrans * (-handleExtent * 10000);
-	// extend the handle a lot further
-	auto q1 = gizmoTrans * (handleExtent * 10000);
+	for (auto i = 0; i<3; i++)
+	{
+		if (handles[i]->isHit(rayPos, rayDir)) {
+			auto hit = handles[i]->getHitPos(rayPos, rayDir);// bad, move hitPos to ref variable
+			auto dist = hitPos.distanceToPoint(rayPos);
+			if (dist < closestDistance) {
+				closestHandle = handles[i];
+				closestDistance = dist;
+				hitPos = hit;
+			}
+		}
+	}
 
-	auto p2 = rayPos;
-	auto q2 = rayPos + rayDir * 1000;
-
-	float s, t;
-	QVector3D c1, c2;
-	auto dist = iris::MathHelper::closestPointBetweenSegments(p1, q1, p2, q2, s, t, c1, c2);
-
-	return c1;
+	return closestHandle;
 }
 
-*/
+void TranslationGizmo::render(QOpenGLFunctions_3_2_Core* gl, QMatrix4x4& viewMatrix, QMatrix4x4& projMatrix)
+{
+	gl->glClear(GL_DEPTH_BUFFER_BIT);
+	//gl->glDisable(GL_DEPTH_TEST);
+	shader->bind();
+
+	shader->setUniformValue("u_viewMatrix", viewMatrix);
+	shader->setUniformValue("u_projMatrix", projMatrix);
+	shader->setUniformValue("showHalf", false);
+
+	for (int i = 0; i<3; i++) {
+		auto transform = this->getTransform();
+		transform.scale(getGizmoScale() * handles[i]->handleRadius);
+		shader->setUniformValue("color", handles[i]->getHandleColor());
+		shader->setUniformValue("u_worldMatrix", transform);
+		//handles[i]->draw(gl, shader);
+		handleMeshes[i]->draw(gl, shader);
+	}
+
+	shader->release();
+}
